code review - fix pipeline wizard hooks
