+++ src/Routes/SidebarRight/AddPipeline/index.js
@@ -1,24 +1,35 @@
-import React from 'react';
+import React, { useState, useReducer, useCallback } from 'react';
+import { addPipelineTemplate } from 'config';
 import { WIZARD_STATE } from 'const';
 import PropTypes from 'prop-types';
 import { useWizardAddPipeline } from 'hooks';
+import { Form } from 'antd';
 import Editor from './Editor';
 import Wizard from './Wizard';
 
 const AddPipeline = ({ jsonPipeline }) => {
-  const {
-    form,
-    isEdit,
-    isEditorVisible,
-    setWizardStepIdx,
-    wizardStepIdx,
-    wizardClear,
-    toggle,
-    handleSubmit,
-    editorState,
-    setEditorState,
+  const [form] = Form.useForm();
+  const [status, setStatus] = useState(WIZARD_STATE.IDLE);
+  const [isEditorVisible, toggle] = useReducer(visible => !visible, false);
+  const [editorState, setEditorState] = useState(addPipelineTemplate);
+  const [wizardStepIdx, setWizardStepIdx] = useState(0);
+  const [isEdit] = useState(jsonPipeline !== undefined);
+
+  const { handleSubmit } = useWizardAddPipeline(
+    jsonPipeline,
     status,
-  } = useWizardAddPipeline(jsonPipeline);
+    isEdit,
+    setEditorState,
+    setStatus,
+    form,
+    setWizardStepIdx,
+    editorState,
+    addPipelineTemplate
+  );
+
+  const wizardClear = useCallback(() => {
+    setStatus(WIZARD_STATE.CLEAR);
+  }, [setStatus]);
 
   if (status === WIZARD_STATE.IDLE) return null;
